Completed AblyAuth unit tests
