+++ src/IO.Ably/AblyRest.cs
@@ -14,8 +14,8 @@ namespace IO.Ably
     /// <summary>Client for the ably rest API</summary>
     public sealed class AblyRest : AblyBase, IRestClient
     {
-        internal AblyHttpClient HttpClient { get; set; }
-        internal MessageHandler MessageHandler;
+        internal AblyHttpClient HttpClient { get; private set; }
+        internal MessageHandler MessageHandler { get; private set; }
 
         internal AblyAuth AblyAuth { get; private set; }
 
@@ -102,7 +102,7 @@ namespace IO.Ably
                 if (ex.ErrorInfo.IsUnAuthorizedError
                     && ex.ErrorInfo.IsTokenError && AblyAuth.TokenRenewable)
                 {
-                    await AblyAuth.Authorise(null, null, true);
+                    await AblyAuth.Authorise(null, new AuthOptions() {Force = true});
                     await AblyAuth.AddAuthHeader(request);
                     return await ExecuteHttpRequest(request);
                 }
