+++ src/IO.Ably/AuthOptions.cs
@@ -34,6 +34,8 @@ namespace IO.Ably
         public string Token { get; set; }
         public TokenDetails TokenDetails { get; set; }
 
+        public bool Force { get; set; }
+
         public Dictionary<string, string> AuthHeaders { get; set; }
         public Dictionary<string, string> AuthParams { get; set; }
         public bool QueryTime { get; set; }
@@ -64,6 +66,8 @@ namespace IO.Ably
         {
             if (AuthCallback == null) AuthCallback = defaults.AuthCallback;
             if (AuthUrl == null) AuthUrl = defaults.AuthUrl;
+            if (Token.IsEmpty()) Token = defaults.Token;
+            if (TokenDetails == null) TokenDetails = defaults.TokenDetails;
             if (AuthHeaders.Count == 0) AuthHeaders = defaults.AuthHeaders;
             if (AuthParams.Count == 0) AuthParams = defaults.AuthParams;
             if (Key.IsEmpty()) Key = defaults.Key;
